+++ src/main/java/com/capgemini/capfoot/entity/Admin.java
@@ -1,8 +1,13 @@
 package com.capgemini.capfoot.entity;
 
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 @Entity
 public class Admin {
-
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
 }
