+++ src/main/java/com/capgemini/capfoot/entity/Groupe.java
@@ -1,7 +1,12 @@
 package com.capgemini.capfoot.entity;
 
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 @Entity
 public class Groupe {
+    @Id @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
 }
