+++ src/main/java/com/capgemini/capfoot/entity/Match.java
@@ -20,7 +20,7 @@ import lombok.NoArgsConstructor;
 public class Match {
 
 	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
 	private boolean phaseDeGroupe;
 	private boolean phaseEliminationDirecte;
