+++ src/main/java/com/capgemini/capfoot/entity/Statut.java
@@ -1,0 +1,5 @@
+package com.capgemini.capfoot.entity;
+
+public enum Statut {
+    INSCRIPTION, GROUPE, QUART_FINAL, DEMI_FINAL, FINAL
+}
