+++ src/main/java/com/capgemini/capfoot/repository/EquipeRepository.java
@@ -1,0 +1,7 @@
+package com.capgemini.capfoot.repository;
+
+import com.capgemini.capfoot.entity.Equipe;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EquipeRepository extends JpaRepository<Equipe, Long> {
+}
