+++ src/main/java/com/capgemini/capfoot/repository/GroupeRepository.java
@@ -1,0 +1,7 @@
+package com.capgemini.capfoot.repository;
+
+import com.capgemini.capfoot.entity.Groupe;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GroupeRepository extends JpaRepository<Groupe,Long> {
+}
