+++ src/main/java/com/capgemini/capfoot/repository/JoueurRepository.java
@@ -1,0 +1,7 @@
+package com.capgemini.capfoot.repository;
+
+import com.capgemini.capfoot.entity.Joueur;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JoueurRepository extends JpaRepository<Joueur, Long> {
+}
